+++ demo/meeting/src/components/ActivityBar/index.tsx
@@ -27,4 +27,6 @@ const ActivityBar = React.forwardRef((props, ref: any) => (
   </Track>
 ));
 
+ActivityBar.displayName = 'MicrophoneActivityBar';
+
 export default ActivityBar;