+++ demo/meeting/src/views/DeviceSetup/index.tsx
@@ -2,14 +2,11 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import React from 'react';
-import {
-  Heading,
-  DeviceSelection,
-} from 'amazon-chime-sdk-component-library-react';
+import { Heading } from 'amazon-chime-sdk-component-library-react';
 
 import JoinMeetingDetails from '../../containers/MeetingJoinDetails';
-
 import { StyledLayout } from './Styled';
+import DeviceSelection from '../../components/DeviceSelection';
 
 const DeviceSetup: React.FC = () => (
   <StyledLayout>